Make Message pickleable, and move transports to System
This commit makes the Message class pickleable, and as part of that moves
the transport pipeline to the System class.

When pickling a Message instance, the pickled message will retain the look
of the original message, however the parent message chain will be squashes
to only include the initial message.